Updating some nugets etc.
