+++ src/Swetugg.Web/Controllers/HomeController.cs
@@ -1,14 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data.Entity.Migrations;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
-using Swetugg.Web.Migrations;
 
 namespace Swetugg.Web.Controllers
 {
-	public class HomeController : Controller
+    public class HomeController : Controller
 	{
 		public ActionResult Index()
 		{
@@ -19,40 +16,6 @@ namespace Swetugg.Web.Controllers
 	    public ActionResult Start()
         {
             return View();
-        }
-
-	    [Route("migration")]
-        [RequireHttps]
-        [Authorize(Roles = "Administrator")]
-	    public ActionResult Migrations(string errorMsg)
-	    {
-            var configuration = new Configuration();
-            var migrator = new DbMigrator(configuration);
-	        ViewBag.AppliedMigrations = migrator.GetDatabaseMigrations();
-	        ViewBag.PendingMigrations = migrator.GetPendingMigrations();
-	        ViewBag.ErrorMessage = errorMsg;
-
-	        return View();
-	    }
-
-        [Route("migration/run")]
-        [RequireHttps]
-        [Authorize(Roles = "Administrator")]
-        public ActionResult RunMigrations()
-        {
-            var configuration = new Configuration();
-            var migrator = new DbMigrator(configuration);
-
-            try
-            {
-                migrator.Update();
-            }
-            catch (Exception ex)
-            {
-                return RedirectToAction("Migrations", new { errorMsg = ex.Message});
-            }
-
-            return RedirectToAction("Migrations");
         }
 
 	}
