+++ src/Swetugg.Web/Views/Database/Migrations.cshtml
@@ -32,7 +32,7 @@
                 }
             </div>
             <div class="panel-footer">
-                @using (Html.BeginForm("RunMigrations", "Home"))
+                @using (Html.BeginForm("RunMigrations", "Database"))
                 {
                     <button class="btn btn-primary @(pendingMigrations.Any() ? "enabled" : "disabled") " type="submit">Apply</button>
                 }
