+++ src/Swetugg.Web/Views/Shared/_Layout.cshtml
@@ -37,7 +37,7 @@
                                         @Html.ActionLink("Users", "Index", "Users", routeValues: new { Area = "Admin" }, htmlAttributes: null)
                                     </li>
                                     <li>
-                                        @Html.ActionLink("Migration", "Migrations", "Home", routeValues: new {Area = ""}, htmlAttributes: null)
+                                        @Html.ActionLink("Migration", "Migrations", "Database", routeValues: new {Area = ""}, htmlAttributes: null)
                                     </li>
                                 }
                             </ul>
