+++ src/Swetugg.Web/packages.config
@@ -5,6 +5,7 @@
   <package id="CommonServiceLocator" version="1.3" targetFramework="net45" />
   <package id="Dapper" version="1.40" targetFramework="net45" />
   <package id="EntityFramework" version="6.1.3" targetFramework="net45" />
+  <package id="Hyak.Common" version="1.0.3" targetFramework="net45" />
   <package id="jQuery" version="2.1.3" targetFramework="net45" />
   <package id="jQuery.Validation" version="1.13.1" targetFramework="net45" />
   <package id="jQuery.Validation.Globalize" version="1.0.1" targetFramework="net45" />
@@ -22,14 +23,18 @@
   <package id="Microsoft.AspNet.Razor" version="3.2.3" targetFramework="net45" />
   <package id="Microsoft.AspNet.Web.Optimization" version="1.1.3" targetFramework="net45" />
   <package id="Microsoft.AspNet.WebPages" version="3.2.3" targetFramework="net45" />
-  <package id="Microsoft.Bcl" version="1.1.8" targetFramework="net45" />
+  <package id="Microsoft.Azure.Common" version="2.1.0" targetFramework="net45" />
+  <package id="Microsoft.Azure.Common.Dependencies" version="1.0.0" targetFramework="net45" />
+  <package id="Microsoft.Azure.KeyVault.Core" version="1.0.0" targetFramework="net45" />
+  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="net45" />
   <package id="Microsoft.Bcl.Async" version="1.0.168" targetFramework="net45" />
-  <package id="Microsoft.Bcl.Build" version="1.0.14" targetFramework="net45" />
-  <package id="Microsoft.Data.Edm" version="5.6.2" targetFramework="net45" />
-  <package id="Microsoft.Data.OData" version="5.6.2" targetFramework="net45" />
-  <package id="Microsoft.Data.Services.Client" version="5.6.2" targetFramework="net45" />
-  <package id="Microsoft.Diagnostics.Tracing.EventSource.Redist" version="1.1.16-beta" targetFramework="net45" />
+  <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="net45" />
+  <package id="Microsoft.Data.Edm" version="5.6.4" targetFramework="net45" />
+  <package id="Microsoft.Data.OData" version="5.6.4" targetFramework="net45" />
+  <package id="Microsoft.Data.Services.Client" version="5.6.4" targetFramework="net45" />
+  <package id="Microsoft.Diagnostics.Tracing.EventSource.Redist" version="1.1.26" targetFramework="net45" />
   <package id="Microsoft.jQuery.Unobtrusive.Validation" version="3.2.3" targetFramework="net45" />
+  <package id="Microsoft.Net.Http" version="2.2.22" targetFramework="net45" />
   <package id="Microsoft.Owin" version="3.0.1" targetFramework="net45" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="3.0.1" targetFramework="net45" />
   <package id="Microsoft.Owin.Security" version="3.0.1" targetFramework="net45" />
@@ -40,18 +45,18 @@
   <package id="Microsoft.Owin.Security.OAuth" version="3.0.1" targetFramework="net45" />
   <package id="Microsoft.Owin.Security.Twitter" version="3.0.1" targetFramework="net45" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net45" />
-  <package id="Microsoft.WindowsAzure.ConfigurationManager" version="3.1.0" targetFramework="net45" />
+  <package id="Microsoft.WindowsAzure.Management.Sql" version="5.2.0" targetFramework="net45" />
   <package id="Modernizr" version="2.8.3" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net45" />
   <package id="Owin" version="1.0" targetFramework="net45" />
   <package id="Respond" version="1.4.2" targetFramework="net45" />
-  <package id="Sendgrid" version="5.1.0" targetFramework="net45" />
-  <package id="SendGrid.SmtpApi" version="1.2.1" targetFramework="net45" />
+  <package id="Sendgrid" version="6.1.0" targetFramework="net45" />
+  <package id="SendGrid.SmtpApi" version="1.3.1" targetFramework="net45" />
   <package id="structuremap" version="3.1.1.134" targetFramework="net45" />
   <package id="StructureMap.MVC5" version="3.1.1.134" targetFramework="net45" />
   <package id="structuremap.web" version="3.1.0.133" targetFramework="net45" />
-  <package id="System.Spatial" version="5.6.2" targetFramework="net45" />
+  <package id="System.Spatial" version="5.6.4" targetFramework="net45" />
   <package id="WebActivatorEx" version="2.0.5" targetFramework="net45" />
   <package id="WebGrease" version="1.6.0" targetFramework="net45" />
-  <package id="WindowsAzure.Storage" version="4.3.0" targetFramework="net45" />
+  <package id="WindowsAzure.Storage" version="6.0.0" targetFramework="net45" />
 </packages>
